- Feature release: SimpleBackup Updated to Release v2.2.5 on the 04/05-2021
- Feature release: SimpleBackup Updated to Release v2.2.5 on the 04/05-2021
- Added checksumverification and integrityTest to 
  Appfunction: varAppFunctionUpdateArchive
  Appfunction: varAppFunctionExtractFilestoFolder
  Appfunction: varAppFunctionExtractFilesWithFullFilePath

- Renamed varBackupSynchronizationDuringBackup to varBackupSynchronization
  It is available in 2 Appfunctions (Backup and update).
  Added missing configuration console output to the update AppFunction.
- Updated documentation.

A note regarding renaming. If you have lots of ini-files it is recommended to use Notepad++ to rename values.
Search-replace in files - Search for whole words and enable case sensitivity.
The following entries have been renamed:
    varChecksumVerificationDuringBackup to varChecksumVerification
    varIntegrityTestDuringBackup to varIntegrityTest
    varBackupSynchronizationDuringBackup to varBackupSynchronization
Please remember to backup your existing ini-files before renaming in case it goes wrong.

It is higly recommended to update to this version.